+++ Lawn/System/ProfileMgr.cpp
@@ -42,8 +42,8 @@ void ProfileMgr::SyncState(DataSync& theSync)
         {
             mProfileMap.clear();
 
-            int aMaxProfileId = 0;
-            int aMaxUseSeq = 0;
+            ulong aMaxProfileId = 0;
+            ulong aMaxUseSeq = 0;
             for (int aProfileCount = aReader->ReadShort(); aProfileCount > 0; aProfileCount--)
             {
                 PlayerInfo aProfile;
